requirements.txt
discord==2.3.2
Flask==3.0.3
python-dotenv==0.21.0
